Improved buy strategy.
Altered the GDAX buy algorithm to weight the purchase by how 'underweight' each product is. In this case, underweight means that the coin's percentage of your portfolio is smaller than the coin's 'relative market cap'. Relative market cap is the coin's market cap relative to the other selected coins to purchase.